Fixed bug when running in React.StrictMode
Elements would be added to the list twice. Was grabbing the tasks
array from this.state.tasks and pushing to it, then returning the
same array in the setState method. Not entirely sure why it caused
it to render twice but calling concat on the previous states object
and reassigning tasks to the new array cleared it up. Calling push
also had another ill effect on shouldComponentUpdate. Since the
address of the current states tasks and nextStates tasks were the
same, you couldn't easily determined if the array would change
after update.